style: drop redundant size_t casts on strong position indices
StrongPositionIndexType<Tag> implicitly converts to Index (int32_t),
which then promotes to size_t for std::vector / std::span subscripts.
The static_cast<std::size_t>(scene_index) scattered throughout sddp_*
was therefore pure noise — dropped in every pure-subscript site.

Comparisons between a strong index and container .size() still trigger
-Wsign-compare, so those sites switch to std::ssize() instead (signed
on both sides of <).  sddp_common.hpp's row.resize(phase_index + 1,…)
uses the next(phase_index) strong-typed helper so no cast is needed
there either.

Drive-by fixes from the cpp26-modernizer review:
- OutputContext::options() forwarding accessor now noexcept
- planning_enums.hpp Doxygen comment: "output_flags" → "write_out"
  (stale from the earlier rename)
- test_output_flags.cpp: std::to_underlying instead of static_cast<uint8_t>

Net -7 lines, 2607/2607 tests still pass.

+++ source/monolithic_method.cpp
@@ -154,7 +154,7 @@ auto MonolithicMethod::solve(PlanningLP& planning_lp, const SolverOptions& opts)
                     .count();
             {
               const std::scoped_lock lk(times_mutex);
-              scene_times[static_cast<std::size_t>(scene_index)] = elapsed;
+              scene_times[scene_index] = elapsed;
             }
             ++scenes_done;
             SPDLOG_INFO("MonolithicMethod: scene {} done in {:.3f}s ({}/{})",
